+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Removed
 - Removed OnCampus section from Calendar settings [#2045](https://github.com/rokwire/illinois-app/issues/2045).
+<<<<<<< HEAD
+### Added
+- Added default attributes from student classification to Analytics events [#2047](https://github.com/rokwire/illinois-app/issues/2047).
+=======
 ### Changed
 - Label to "My Gies Canvas Courses" [#2048](https://github.com/rokwire/illinois-app/issues/2048).
+>>>>>>> develop
 
 ## [4.0.54] - 2022-07-18
 ### Added